feat: Add battle character evaluator for white list attribute
This commit adds a battle character evaluator for white list attribute. It includes the implementation of the IBattleCharacterEvaluator interface in the AddAilment.cs file. The evaluator allows for evaluating a battle character based on a white list of character attributes. This addition enhances the flexibility of character attribute evaluation in the game.

Co-authored-by: [Author Name]