feat: Add ShowMessageData struct for displaying messages
The commit adds a new file `ShowMessageData.cs` which contains a `ShowMessageData` struct. This struct has two properties: `Message` and `SfxName`. It also has a constructor that initializes these properties. This struct will be used for displaying messages in the game.

Note: The commit message has been generated based on the provided code changes and recent commits.