feat: Update message display in ExplorationView and OnRequestShowMessage
This commit modifies the code in ExplorationView.cs and OnRequestShowMessage.cs to update how messages are displayed in the game. In ExplorationView.cs, the CreateElement method now takes the message as a parameter instead of the entire object. In OnRequestShowMessage.cs, the ShowMessageAndWaitForSubmitInputAsync method now takes a ShowMessageData object instead of a string.

These changes improve the flexibility and reusability of the message display functionality in the game.