feat: Add default constructor to ShowMessageData class
This commit adds a default constructor to the `ShowMessageData` class in the `ShowMessageData.cs` file. The default constructor initializes the `Message` property to an empty string and the `SfxName` property to `null`. This allows for easier instantiation of `ShowMessageData` objects without having to provide message and sound effect parameters.

Note: The commit message has been generated based on the provided code changes and recent commits.