feat: Add OnRequestPlaySfx event to GameEvents.cs
This commit adds a new event, OnRequestPlaySfx, to the GameEvents.cs file. The event is of type Subject<string> and is used to request the playing of sound effects in the game. This event will be subscribed to in the GameSceneController.cs and ExplorationView.cs files to play sound effects when certain actions occur.

Note: The commit message has been generated based on the provided code changes and recent commits.