feat: Add sfxName parameter to ShowMessageAndWaitForSubmitInputAsync
This commit adds a new parameter, sfxName, to the ShowMessageAndWaitForSubmitInputAsync method in the GameEvents.cs file. The sfxName parameter allows for specifying the name of the sound effect to be played when showing the message. This change enhances the audio experience in the game.

Note: The commit message has been generated based on the provided code changes and recent commits.