feat: Fix default value for sfxName parameter in GiveDamage.cs
This commit fixes the default value for the sfxName parameter in the GiveDamage.cs file. Previously, the default value was not set correctly, causing issues with playing the sound effect when showing the message. By setting the default value to "Sfx.Message.1", the sound effect will now play correctly, enhancing the audio experience in the game.