feat: Add TakeDamage method to BattleCharacter class
This commit adds a new method, TakeDamage, to the BattleCharacter class in the BattleCharacter.cs file. The TakeDamage method is responsible for deducting the specified amount of damage from the character's BattleStatus and triggering the OnTakeDamage event. This addition enhances the functionality of the BattleCharacter class by allowing it to receive and handle damage.

Note: The commit message has been generated based on the provided code changes and recent commits.