feat: Add HKUIDocument field to GameCameraController
This commit adds a new HKUIDocument field to the GameCameraController class in the GameCameraController.cs file. The HKUIDocument is used to reference a UI document for playing animations asynchronously. This addition enhances the functionality of the GameCameraController by allowing it to play animations based on game events.

Note: The commit message has been generated based on the provided code changes and recent commits.