feat: Update MasterData.cs to add NeedBehaviourPoint and NeedStamina properties
This commit updates the MasterData.cs file to add the NeedBehaviourPoint and NeedStamina properties. These properties are used to determine the behavior point and stamina requirements for a character. This enhancement improves the functionality of the MasterData class by providing a way to track and manage the behavior point and stamina needs of characters.

Note: The commit message has been generated based on the provided code changes and recent commits.